refactor: chenge restructur file sistem
