+++ package.json
@@ -2,7 +2,7 @@
   "name": "task_1_cli_cipher",
   "version": "1.0.0",
   "description": "CLI_Task1 консольное приложение позволяющее производить шифрование или дешифрование текстовых файлов, а также вводимого в консоли текста в соответствии с заданными алгоритмами шифрования.",
-  "main": "a.js",
+  "main": "app.js",
   "scripts": {
     "test": "jest",
     "test:coverage": "jest --coverage"
